improve eclp input stability
