+++ packages/nextjs/app/v3/_components/ChooseParameters/NumberParameterButton.tsx
@@ -1,0 +1,23 @@
+export function NumberParameterButton({
+  value,
+  selectedValue,
+  onClick,
+  isPercentage = false,
+}: {
+  isPercentage?: boolean;
+  value: string;
+  selectedValue: string;
+  onClick: () => void;
+}) {
+  return (
+    <div
+      className={`bg-base-200 w-20 h-12 rounded-lg flex items-center justify-center hover:cursor-pointer hover:border-2 hover:border-accent text-lg ${
+        value === selectedValue ? "border-2 border-accent" : ""
+      }`}
+      onClick={onClick}
+    >
+      {value}
+      {isPercentage && "%"}
+    </div>
+  );
+}
