+++ packages/nextjs/app/v3/_components/ChooseParameters/types.ts
@@ -1,0 +1,6 @@
+export type HandleNumberInputChange = (
+  e: React.ChangeEvent<HTMLInputElement>,
+  field: "swapFeePercentage" | "amplificationParameter",
+  min: number,
+  max: number,
+) => void;
